changed how the find an replace works
