+++ srcs/ft_unsetenv.c
@@ -45,86 +45,3 @@ void	ft_unsetenv(char **env, char *arg)
 		ft_putendl("Variable not found");
 	ft_pushback(env, &len);
 }
-
-char	*ft_find_variable(char *path)
-{
-	char *tmp;
-	char **tab;
-	char *ret;
-
-	tmp = ft_strstr(path, "$");
-	tab = ft_strsplit(tmp, '/');
-	tmp = ft_strdup(tab[0]);
-	ft_tabfree(tab);
-	ret = ft_strstr(tmp, &tmp[1]);
-	return ret;
-}
-
-char	*ft_getvar(char *path, char **env)
-{
-	char *var;
-	int i;
-	size_t len;
-	char *ret;
-	
-	var = ft_find_variable(path);
-	i = 0;
-	len = ft_strlen(var);
-	while (env[i])
-	{
-		if (ft_strncmp(env[i], var, len) == 0)
-		{	
-			ret = ft_strstr(env[i], "=");
-			var = ft_strstr(ret, &ret[1]);
-			return(var);
-		}
-		i++;
-	}
-	return NULL;
-}
-
-char	**ft_substitute(char *path, char **env)
-{
-	int i;
-	char *var;
-	char *value;
-	char **tab;
-
-
-	i = 0;
-	var = ft_strjoin("$", ft_find_variable(path));
-	value = ft_getvar(path, env);
-	tab = ft_strsplit(path, '/');
-	while (tab[i])
-	{
-		if (ft_strcmp(tab[i], var) == 0){
-			ft_strdel(&tab[i]);
-			tab[i] = ft_strdup(value);	
-		}
-		i++;
-	}
-	return tab;
-}
-
-char  *handler(char *path, char **env)
-{
-
-	return(ft_newpath(ft_substitute(path, env)));
-}
-
-char *ft_newpath(char **tab)
-{
-	int i;
-	char *tmp;
-	
-	i = 0;
-	tmp = "/";
-	while(tab[i] != '\0')
-	{	
-		tmp = ft_strjoin(tmp, tab[i]);
-		tmp = ft_strjoin(tmp, "/");
-		i++;
-	}
-	return tmp;
-}
-
